fix(svelte): update test component imports for renamed stream.custom.svelte.ts
